added new select sign, gun in front, updated arcade sounds to 2d, duck sounds better 3d
