--- huntduck/Assets/HealthBar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ef33532064e04e0c9475871b8c1ce89
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
