--- huntduck/Assets/PlayerHealthUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cb516887b04d5490a8f5c0bbacca9feb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
