+++ huntduck/Assets/Scripts/SelectManager.cs
@@ -7,42 +7,35 @@ public class SelectManager : MonoBehaviour
 {
     [Header("UI")]
     public GameObject welcomeUI;
+    public GameObject selectGunUI;
+    public Text selectGunText;
     public GameObject helperUI;
     public Text helperText;
     public GameObject buttons;
+    public GameObject downArrow;
 
     [Header("Weapons")]
     public GameObject weaponsWall;
 
     private void OnEnable()
     {
-        SelectModeTrigger.onSelectModeTriggered += StartSelectMode;
-        WallSlider.onPosition1Reached += ShowWeaponsWall;
+        StartSelectMode();
+        WeaponsManager.onWeaponSelected += HideSelectUI;
         WeaponsManager.onWeaponSelected += ShowButtons;
+        WallSlider.onPosition1Reached += ShowWeaponsWall;
     }
 
     private void OnDisable()
     {
+        WeaponsManager.onWeaponSelected -= HideSelectUI;
         WeaponsManager.onWeaponSelected -= ShowButtons;
         WallSlider.onPosition1Reached -= ShowWeaponsWall;
-        SelectModeTrigger.onSelectModeTriggered -= StartSelectMode;
-    }
-
-    void ShowButtons()
-    {
-        helperText.text = "SHOOT PRACTICE OR HUNT TO BEGIN";
-        // turn on button gameobject
-        buttons.SetActive(true);
+        downArrow.SetActive(false); // hide this arrow on sign, activate by script when needed
     }
 
     void StartSelectMode()
     {
         StartCoroutine(SelectModeIntro());
-    }
-
-    void ShowWeaponsWall()
-    {
-        weaponsWall.SetActive(true);
     }
 
     IEnumerator SelectModeIntro()
@@ -51,7 +44,25 @@ public class SelectManager : MonoBehaviour
         yield return new WaitForSeconds(3f);
         welcomeUI.SetActive(false);
 
+        selectGunUI.SetActive(true);
+        selectGunText.text = "SELECT YOUR WEAPON \n "; // give space for down arrow
+        downArrow.SetActive(true);
+    }
+
+    void HideSelectUI()
+    {
+        selectGunUI.SetActive(false);
+    }
+
+    void ShowButtons()
+    {
         helperUI.SetActive(true);
-        helperText.text = "SELECT YOUR WEAPON TO BEGIN";
+        helperText.text = "SHOOT PRACTICE OR HUNT TO BEGIN";
+        buttons.SetActive(true);
+    }
+
+    void ShowWeaponsWall()
+    {
+        weaponsWall.SetActive(true);
     }
 }
