+++ huntduck/Assets/Sounds/Weapons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f945d55938eaa43f58532b139a25af52
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
