+++ huntduck/Assets/Sounds/Weapons/zapsplat_multimedia_game_sound_laser_gun_fire_single_shot_retro_001_79071.mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: e0b269037530144b18e2edb0237cf8b4
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
